hgi/requirements.txt
asgiref==3.6.0
Command==0.1.0
Django==3.2.16
djangorestframework==3.14.0
drf-flex-fields==1.0.0
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
typing_extensions==4.4.0
